Resolves #158: Factions as graph (#160)
Welcome to the world of JGraphT :)

The relations between factions are now internally held as a `Graph`. Every newly created Faction registers itself so that the `FactionRelations` object always returns valid values.

The intention of the `graph` module is to serve as THE single module to provide the JGraphT dependency. It will also contain convenience methods to handle graphs more "kotlin style".